Fix tests from stringformatter location refactor
